This function computes loads in a set of promoter capture Hi-C files (pcHi-C). This is a tab delimited file that contains the following mandatory columns

ensg - Ensembl gene id for gene

name - Gene name

biotype - Gene biotype from ensembl e.g. protein_coding

strand - Gene strand

baitChr - Bait (A captured restriction fragment) chromosome

baitStart - Bait fragment start

baitEnd - Bait fragment end

baitID - Bait fragment ID should match ID in digest file

baitName - Free text of what is captured - deprecated.

oeChr - 'Other end' (Promoter interacting restriction fragment) chromosome

oeStart - 'Other end' fragment start

oeEnd - 'Other end' fragment end

oeID - 'Other end' fragment ID should match ID in digest file

oeName - Free text of what is interacting - deprecated.

Additional columns contain CHiCAGO scores for one or more columns.

This function computes loads in a set of promoter capture Hi-C files (pcHi-C). This is a tab delimited file that contains the following mandatory columns

ensg - Ensembl gene id for gene

name - Gene name

biotype - Gene biotype from ensembl e.g. protein_coding

strand - Gene strand

baitChr - Bait (A captured restriction fragment) chromosome

baitStart - Bait fragment start

baitEnd - Bait fragment end

baitID - Bait fragment ID should match ID in digest file

baitName - Free text of what is captured - deprecated.

oeChr - 'Other end' (Promoter interacting restriction fragment) chromosome

oeStart - 'Other end' fragment start

oeEnd - 'Other end' fragment end

oeID - 'Other end' fragment ID should match ID in digest file

oeName - Free text of what is interacting - deprecated.

Additional columns contain CHiCAGO scores for one or more columns.

make_pchic(f, chic_thresh = 5, biotype.filter = "protein_coding",
  remove.trans.contacts = TRUE)

Arguments

f
a scalar representing file name of pcHi-C file

chic_thresh
scalar representing CHiCAGO threshold for calling a significant interaction. Default is set to 5.

biotype.filter
scalar/vector of characters representing a biotype filter. Default is set to 'protein_coding'

remove.trans.contacts
a boolean on whether to remove interaction between chromosomes. Default is set to TRUE.

Value

list of data.tables where each data.table has three columns ensg,fragid and chic.score